+++ packages/lumx-vue/src/index.ts
@@ -7,6 +7,7 @@ export * from './components/button';
 export * from './components/checkbox';
 export * from './components/chip';
 export * from './components/divider';
+export * from './components/drag-handle';
 export * from './components/flag';
 export * from './components/flex-box';
 export * from './components/grid-column';